Simplified file conversion method, moved it to RepairManager and added error message
