+++ modules/Movecraft/src/main/resources/localisation/movecraftlang_en.properties
@@ -214,7 +214,7 @@ Repair\ functionality\ is\ disabled\ or\ WorldEdit\ was\ not\ detected=Repair fu
 You\ must\ be\ piloting\ a\ craft=You must be piloting a craft
 BLOCK\ IS\ PART\ OF\ A\ PILOTED\ CRAFT=BLOCK IS PART OF A PILOTED CRAFT
 Repait\ -\ Save\ Failed\ Console=Factory save failed, this is typically caused by FAWE
-Repair\ -\ Illegal\ character\ found= ERROR: Illegal character found on this Repair sign:
+Repair\ -\ Invalid\ Player\ Name=The owner of repair state named %s has either changed name or never joined the server\!
 Could\ not\ save\ file=Could not save file
 State\ saved=State saved
 REPAIR\ STATE\ NOT\ FOUND=REPAIR STATE NOT FOUND
