+++ modules/api/src/main/java/net/countercraft/movecraft/MovecraftRepair.java
@@ -32,7 +32,6 @@ import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.block.Dispenser;
 import org.bukkit.block.Sign;
-import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.Plugin;
 
@@ -487,67 +486,5 @@ public class MovecraftRepair {
         return instance;
     }
 
-    public int convertOldCraftRepairStates(){
-        //Check in the old RepairStates folder
-        File repairStateDir = new File(plugin.getDataFolder(),"RepairStates");
-        if (!repairStateDir.exists()){
-            return 0;
-        }
-        int convertedRepairStates = 0;
-        File[] repairStates = repairStateDir.listFiles();
-        if (repairStateDir.exists() && (repairStates != null || repairStates.length > 0)){
-            for (File rs : repairStates){
-                if (!rs.getName().contains(".schematic")){
-                    continue;
-                }
-                String fileName = rs.getName();
-                fileName = fileName.replace(".schematic", "");
-                OfflinePlayer owner = null;
-                if (fileName.contains("_")) {
-                    String[] parts = fileName.split("_");
-                    owner = Bukkit.getPlayer(parts[0]) == null ? Bukkit.getOfflinePlayer(parts[0]) : Bukkit.getPlayer(parts[0]);
-                    fileName = "";
-                    if (parts.length > 2){
-                        for (int i = 1 ; i <= parts.length - 1 ; i++){
-                            fileName += parts[i];
-                            if (i < parts.length - 1){
-                                fileName += "_";
-                            }
-                        }
-                    } else {
-                        fileName = parts[1];
-                    }
-                }
 
-                for (Player p : Bukkit.getOnlinePlayers()){
-                    if (fileName.startsWith(p.getName())){
-                        owner = p;
-                        fileName = fileName.replace(p.getName(),"");
-                        break;
-                    }
-                }
-                if (owner == null){
-                    for (OfflinePlayer op : Bukkit.getOfflinePlayers()){
-                        if (fileName.startsWith(op.getName())){
-                            owner = op;
-                            fileName = fileName.replace(op.getName(),"");
-                        }
-                    }
-                }
-                if (owner == null){
-                    continue;
-                }
-                File playerDir = new File(repairStateDir, owner.getUniqueId().toString());
-                if (!playerDir.exists()) {
-                    playerDir.mkdirs();
-                }
-                File dest = new File(playerDir, fileName + ".schematic");
-                if (rs.renameTo(dest)){
-                    convertedRepairStates++;
-                }
-
-            }
-        }
-        return convertedRepairStates;
-    }
 }
